web-converter-backend/pyproject.toml
[tool.poetry]
name = "web-converter-backend"
version = "0.1.0"
description = ""
authors = ["Ivan Nikolaev-Axenov <nikolaevaxenov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlparse = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
